relying party views separated in more files (#173)
* refactor:  relying_party views seprataed in multiple files

* docs: added advanced entity listing endpoint in AUTHORITY.md

* docs:  auth code redirect and logout endpoints

* docs: minor changes

* docs: specified content type response advanced listing endpoint
