+++ spid_cie_oidc/relying_party/tests/test_04_rp_callback.py
@@ -59,7 +59,7 @@ class RpCallBack(TestCase):
         )
 
     @override_settings(HTTP_CLIENT_SYNC=True)
-    @patch("spid_cie_oidc.relying_party.views.process_user_attributes", return_value=None)
+    @patch("spid_cie_oidc.relying_party.views.rp_callback.process_user_attributes", return_value=None)
     @patch("requests.post", return_value=MockedTokenEndPointResponse())
     @patch("requests.get", return_value=MockedUserInfoResponse())
     def test_rp_callback_no_rp_attr_map(self, mocked, mocked_2, mocked_3):
@@ -69,7 +69,7 @@ class RpCallBack(TestCase):
         self.assertTrue(res.status_code == 403)
 
     @override_settings(HTTP_CLIENT_SYNC=True)
-    @patch("spid_cie_oidc.relying_party.views.process_user_attributes", return_value=None)
+    @patch("spid_cie_oidc.relying_party.views.rp_callback.process_user_attributes", return_value=None)
     @patch("requests.post", return_value=MockedTokenEndPointResponse())
     @patch("requests.get", return_value=MockedUserInfoResponse())
     def test_rp_callback_incorret_request(self, mocked, mocked_2, mocked_3):
@@ -79,9 +79,9 @@ class RpCallBack(TestCase):
         self.assertTrue("error" in res.json())
 
     @override_settings(HTTP_CLIENT_SYNC=True)
-    @patch("spid_cie_oidc.relying_party.views.process_user_attributes", return_value=None)
+    @patch("spid_cie_oidc.relying_party.views.rp_callback.process_user_attributes", return_value=None)
     @patch("requests.post", return_value=MockedTokenEndPointResponse())
-    @patch("spid_cie_oidc.relying_party.views.SpidCieOidcRpCallbackView.get_userinfo", return_value=None)
+    @patch("spid_cie_oidc.relying_party.views.rp_callback.SpidCieOidcRpCallbackView.get_userinfo", return_value=None)
     def test_rp_callback_no_userinfo(self, mocked, mocked_2, mocked_3):
         client = Client()
         url = reverse("spid_cie_rp_callback")
@@ -89,8 +89,8 @@ class RpCallBack(TestCase):
         self.assertTrue(res.status_code == 400)
         
     @override_settings(HTTP_CLIENT_SYNC=True)
-    @patch("spid_cie_oidc.relying_party.views.process_user_attributes", return_value=None)
-    @patch("spid_cie_oidc.relying_party.views.SpidCieOidcRpCallbackView.access_token_request", return_value=None)
+    @patch("spid_cie_oidc.relying_party.views.rp_callback.process_user_attributes", return_value=None)
+    @patch("spid_cie_oidc.relying_party.views.rp_callback.SpidCieOidcRpCallbackView.access_token_request", return_value=None)
     def test_rp_callback_no_token_response(self, mocked, mocked_2):
         client = Client()
         url = reverse("spid_cie_rp_callback")
