+++ spid_cie_oidc/relying_party/views/rp_callback_echo_attributes.py
@@ -1,0 +1,15 @@
+import logging
+
+from django.shortcuts import render
+from django.views import View
+
+
+logger = logging.getLogger(__name__)
+
+
+class SpidCieOidcRpCallbackEchoAttributes(View):
+    template = "echo_attributes.html"
+
+    def get(self, request):
+        data = {"oidc_rp_user_attrs": request.session.get("oidc_rp_user_attrs", {})}
+        return render(request, self.template, data)
